+++ src/main/java/org/example/hospital_admission_project/entity/Rating.java
@@ -14,10 +14,10 @@ import java.util.List;
 public class Rating {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     @ElementCollection
-    private List<Long> usersId;
-    private Long ownerId;
+    private List<Integer> usersId;
+    private Integer ownerId;
     private Double ownerRating;
     @ElementCollection
     private List<Double> ratings = new ArrayList<>();
