+++ src/main/java/org/example/hospital_admission_project/repo/RatingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RatingRepository extends JpaRepository<Rating, Long> {
-    Optional<Rating> findByOwnerId(Long doctorId);
+public interface RatingRepository extends JpaRepository<Rating, Integer> {
+    Optional<Rating> findByOwnerId(Integer doctorId);
 }
