+++ src/main/java/org/example/hospital_admission_project/service/AppointmentService.java
@@ -25,7 +25,7 @@ public class AppointmentService {
     private final ConsultationRepository consultationRepository;
     private final UserService userService;
     private final BookingRepository bookingRepository;
-    public ResponseEntity<?> createBooking(Long doctorId, String reason) {
+    public ResponseEntity<?> createBooking(Integer doctorId, String reason) {
         Optional<Doctor> optionalDoctor = doctorRepository.findById(doctorId);
         if (optionalDoctor.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new SendMessage(false, "Doctor not found", null));
@@ -76,7 +76,7 @@ public class AppointmentService {
     }
 
     public ResponseEntity<?> cancelBooking(Integer bookingId) {
-        Optional<Booking> optionalBooking = bookingRepository.findById(Long.valueOf(bookingId));
+        Optional<Booking> optionalBooking = bookingRepository.findById(bookingId);
 
         if (optionalBooking.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new SendMessage(false, "Booking not found", null));
